Improve performance of Operation.get_attr somewhat
This function is the fourth most expensive in a simple example I have been benchmarking. It had a somewhat odd loop over possible fields when the code has already checked which field is present using WhichOneof. I've fixed this (though there is still a loop when the oneof field was a list, which I believe is unavoidable.)

PiperOrigin-RevId: 223344446